acl update

+++ src/api/v1/controller/acl/RoleGroupController.ts
@@ -5,41 +5,42 @@ import {ValidationError} from "vesta-schema/error/ValidationError";
 import {RoleGroup, IRoleGroup} from "../../../../cmn/models/RoleGroup";
 import {IUpsertResult} from "vesta-schema/ICRUDResult";
 import {Vql} from "vesta-schema/Vql";
+import {Permission} from "../../../../cmn/models/Permission";
 
 
 export class RoleGroupController extends BaseController {
 
-    public route(router:Router) {
-        router.get('/roleGroup/:id', this.checkAcl('acl.roleGroup', 'read'), this.getRoleGroup.bind(this));
-        router.get('/roleGroup', this.checkAcl('acl.roleGroup', 'read'), this.getRoleGroups.bind(this));
-        router.post('/roleGroup', this.checkAcl('acl.roleGroup', 'create'), this.addRoleGroup.bind(this));
-        router.put('/roleGroup', this.checkAcl('acl.roleGroup', 'update'), this.updateRoleGroup.bind(this));
-        router.delete('/roleGroup', this.checkAcl('acl.roleGroup', 'delete'), this.removeRoleGroup.bind(this));
+    public route(router: Router) {
+        router.get('/acl/roleGroup/:id', this.checkAcl('acl.roleGroup', Permission.Action.Read), this.getRoleGroup.bind(this));
+        router.get('/acl/roleGroup', this.checkAcl('acl.roleGroup', Permission.Action.Read), this.getRoleGroups.bind(this));
+        router.post('/acl/roleGroup', this.checkAcl('acl.roleGroup', Permission.Action.Add), this.addRoleGroup.bind(this));
+        router.put('/acl/roleGroup', this.checkAcl('acl.roleGroup', Permission.Action.Edit), this.updateRoleGroup.bind(this));
+        router.delete('/acl/roleGroup/:id', this.checkAcl('acl.roleGroup', Permission.Action.Delete), this.removeRoleGroup.bind(this));
     }
 
     protected init() {
 
     }
 
-    public getRoleGroup(req:IExtRequest, res:Response, next:Function) {
+    public getRoleGroup(req: IExtRequest, res: Response, next: Function) {
         RoleGroup.findById<IRoleGroup>(req.params.id, {relations: ['roles']})
             .then(result=> res.json(result))
             .catch(reason=> this.handleError(res, Err.Code.DBQuery, reason.error.message));
     }
 
-    public getRoleGroups(req:IExtRequest, res:Response, next:Function) {
+    public getRoleGroups(req: IExtRequest, res: Response, next: Function) {
         var query = new Vql('RoleGroup');
-        query.filter(req.params.query);
+        query.filter(req.query.query);
         RoleGroup.findByQuery(query)
             .then(result=>res.json(result))
             .catch(reason=> this.handleError(res, Err.Code.DBQuery, reason.error.message));
     }
 
-    public addRoleGroup(req:IExtRequest, res:Response, next:Function) {
+    public addRoleGroup(req: IExtRequest, res: Response, next: Function) {
         var roleGroup = new RoleGroup(req.body),
             validationError = roleGroup.validate();
         if (validationError) {
-            var result:IUpsertResult<IRoleGroup> = <IUpsertResult<IRoleGroup>>{};
+            var result: IUpsertResult<IRoleGroup> = <IUpsertResult<IRoleGroup>>{};
             result.error = new ValidationError(validationError);
             this.acl.initAcl();
             return res.json(result);
@@ -49,11 +50,11 @@ export class RoleGroupController extends BaseController {
             .catch(reason=> this.handleError(res, Err.Code.DBInsert, reason.error.message));
     }
 
-    public updateRoleGroup(req:IExtRequest, res:Response, next:Function) {
+    public updateRoleGroup(req: IExtRequest, res: Response, next: Function) {
         var roleGroup = new RoleGroup(req.body),
             validationError = roleGroup.validate();
         if (validationError) {
-            var result:IUpsertResult<IRoleGroup> = <IUpsertResult<IRoleGroup>>{};
+            var result: IUpsertResult<IRoleGroup> = <IUpsertResult<IRoleGroup>>{};
             result.error = new ValidationError(validationError);
             return res.json(result);
         }
@@ -68,13 +69,24 @@ export class RoleGroupController extends BaseController {
             .catch(reason=> this.handleError(res, Err.Code.DBUpdate, reason.error.message));
     }
 
-    public removeRoleGroup(req:IExtRequest, res:Response, next:Function) {
-        var roleGroup = new RoleGroup({id: req.body.id});
-        roleGroup.delete()
-            .then(result=>{
-                this.acl.initAcl();
-                res.json(result);
+    public removeRoleGroup(req: IExtRequest, res: Response, next: Function) {
+        RoleGroup.findById(+req.params.id)
+            .then(result=> {
+                if (!result.items.length) return this .handleError(res, Err.Code.DBQuery, 'Record not found');
+                let roleGroupName = result.items[0].name;
+                let security = this.setting.security;
+                if (roleGroupName == security.rootRoleName || roleGroupName == security.guestRoleName) {
+                    return this.handleError(res, Err.Code.Forbidden, 'admin and guest roleGroups are required');
+                }
+                var roleGroup = new RoleGroup({id: +req.params.id});
+                roleGroup.delete()
+                    .then(result=> {
+                        this.acl.initAcl();
+                        res.json(result);
+                    })
+                    .catch(reason=> this.handleError(res, Err.Code.DBDelete, reason.error.message));
             })
-            .catch(reason=> this.handleError(res, Err.Code.DBDelete, reason.error.message));
+            .catch(reason=> this.handleError(res, Err.Code.DBQuery, reason.error.message));
+
     }
 }
